Add input field for facebook id.
- Add button to facebook page.
- Put facebook page form in separate file.
