+++ crowdgezwitscher/facebook/models.py
@@ -23,6 +23,9 @@ class FacebookPage(MapObject):
     def get_absolute_url(self):
         return reverse('facebook:detail', kwargs={'pk': self.pk})
 
+    def url(self):
+        return 'https://www.facebook.com/%s' % self.facebook_id
+
 
 @python_2_unicode_compatible
 class FacebookLikeStatistic(models.Model):
