+++ crowdgezwitscher/facebook/templates/facebook/form.html
@@ -32,6 +32,11 @@
         </div>
     </div>
     <div class="row">
+        <div class="fieldWrapper col-xs-12 col-sm-6 col-md-4 col-lg-4">
+            {% bootstrap_field form.facebook_id %}
+        </div>
+    </div>
+    <div class="row">
         <div class="fieldWrapper col-xs-12 col-sm-12 col-md-12 col-lg-12">
             {% bootstrap_field form.notes %}
         </div>
