+++ crowdgezwitscher/facebook/views.py
@@ -1,36 +1,15 @@
 from rest_framework import generics
 
-from django import forms
 from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.views.generic.list import ListView
 from django.views.generic.detail import DetailView
 from django.views.generic.edit import CreateView, UpdateView, DeleteView
 from django.urls import reverse_lazy
 
-from base.fields import RoundingDecimalField
 from base.models import MapObjectFilterBackend
-from base.widgets import SelectizeSelectMultiple
+from facebook.forms import FacebookPageForm
 from facebook.models import FacebookPage
 from facebook.serializers import FacebookPageSerializer, FacebookPageSerializerShortened
-
-
-class FacebookPageForm(forms.ModelForm):
-    location_lat = RoundingDecimalField(
-        max_digits=9, decimal_places=6, widget=forms.NumberInput(attrs={'class': 'form-control', 'step': 'any'}),
-    )
-    location_long = RoundingDecimalField(
-        max_digits=9, decimal_places=6, widget=forms.NumberInput(attrs={'class': 'form-control', 'step': 'any'}),
-    )
-
-    class Meta:
-        model = FacebookPage
-        fields = ('name', 'active', 'location_long', 'location_lat', 'location', 'notes', 'events')
-        widgets = {
-            'events': SelectizeSelectMultiple(),
-            'location': forms.TextInput(attrs={'class': 'form-control'}),
-            'name': forms.TextInput(attrs={'class': 'form-control'}),
-            'notes': forms.Textarea(attrs={'class': 'form-control', 'rows': 10}),
-        }
 
 
 class FacebookPageListView(PermissionRequiredMixin, ListView):
